etna: Split ShaderProgram to ShaderProgram.hpp and DescriptorSetLayout.hpp
- Files splitted and other small changes
