+++ etna/CMakeLists.txt
@@ -4,7 +4,8 @@ cmake_minimum_required(VERSION 3.20)
 add_library(etna
 	"source/UniqueImage.cpp"
  	"source/VmaImplementation.cpp"
- 	"source/ShaderProgram.cpp")
+ 	"source/ShaderProgram.cpp"
+	"source/DescriptorSetLayout.cpp")
 
 target_include_directories(etna PUBLIC include)
 target_include_directories(etna PRIVATE source)
