+++ etna/include/etna/ShaderProgram.hpp
@@ -3,8 +3,6 @@
 #define ETNA_SHADER_PROGRAM_HPP_INCLUDED
 
 #include <vulkan/vulkan.hpp>
-#include <SPIRV-Reflect/spirv_reflect.h>
-#include <bitset>
 
 #include <array>
 #include <bitset>
@@ -12,68 +10,10 @@
 #include <unordered_map>
 #include <memory>
 
+#include "DescriptorSetLayout.hpp"
+
 namespace etna
 {
-  constexpr uint32_t MAX_PROGRAM_DESCRIPTORS = 4u;
-  constexpr uint32_t MAX_DESCRIPTOR_BINDINGS = 8u; /*If you are out of bindings, try using arrays of images/samplers*/
-  
-  struct DescriptorSetLayoutHash;
-
-  struct DescriptorSetInfo
-  {
-    void parseShader(vk::ShaderStageFlagBits stage, const SpvReflectDescriptorSet &spv);
-    void addResource(const vk::DescriptorSetLayoutBinding &binding);
-    void merge(const DescriptorSetInfo &info);
-
-    bool operator==(const DescriptorSetInfo &rhs) const;
-
-    vk::DescriptorSetLayout createVkLayout(vk::Device device) const;
-
-    void clear();
-  private:
-    uint32_t maxUsedBinding = 0;
-    uint32_t dynOffsets = 0;
-
-    std::bitset<MAX_DESCRIPTOR_BINDINGS> usedBindings {};
-    std::array<vk::DescriptorSetLayoutBinding, MAX_DESCRIPTOR_BINDINGS> bindings {};
-
-    friend DescriptorSetLayoutHash;
-  };
-
-  struct DescriptorSetLayoutHash
-  {
-    std::size_t operator()(const DescriptorSetInfo &res) const;
-  };
-
-  using DescriptorLayoutId = uint32_t;
-
-  struct DescriptorSetLayoutCache
-  {
-    DescriptorSetLayoutCache() {}
-    ~DescriptorSetLayoutCache() { /*make device global and call clear hear*/}
-    
-    DescriptorLayoutId registerLayout(vk::Device device, const DescriptorSetInfo &info);
-    void clear(vk::Device device);
-
-    const DescriptorSetInfo &getLayoutInfo(DescriptorLayoutId id) const
-    {
-      return descriptors.at(id);
-    }
-
-    VkDescriptorSetLayout getVkLayout(DescriptorLayoutId id) const
-    {
-      return vkLayouts.at(id);
-    }
-
-    DescriptorSetLayoutCache(const DescriptorSetLayoutCache&) = delete;
-    DescriptorSetLayoutCache &operator=(const DescriptorSetLayoutCache&) = delete; 
-
-  private:
-    std::unordered_map<DescriptorSetInfo, DescriptorLayoutId, DescriptorSetLayoutHash> map;
-    std::vector<DescriptorSetInfo> descriptors;
-    std::vector<vk::DescriptorSetLayout> vkLayouts;
-  };
-
   struct ShaderModule
   {
     ShaderModule(vk::Device device, const std::string &shader_path);
@@ -86,8 +26,6 @@ namespace etna
     vk::ShaderStageFlagBits getStage() const { return stage; }
     const std::string &getName() const { return entryPoint; }
     vk::PushConstantRange getPushConst() const { return pushConst; }
-
-    bool getPushConstUsed() const { return pushConst.size !=0u; }
 
     ShaderModule(const ShaderModule &mod) = delete;
     ShaderModule &operator=(const ShaderModule &mod) = delete;  
@@ -143,6 +81,8 @@ namespace etna
 
     struct ShaderProgramInternal
     {
+      ShaderProgramInternal(std::vector<uint32_t> &&mod) : moduleIds {std::move(mod)} {}
+
       std::vector<uint32_t> moduleIds;
 
       std::bitset<MAX_PROGRAM_DESCRIPTORS> usedDescriptors;
@@ -150,14 +90,13 @@ namespace etna
     
       vk::PushConstantRange pushConst {};
       vk::PipelineLayout progLayout {nullptr};
+
+      void reload(ShaderProgramManager &manager);
+      void destroy(ShaderProgramManager &manager);
     };
 
     std::unordered_map<std::string, uint32_t> programNames;
     std::vector<std::unique_ptr<ShaderProgramInternal>> programs;
-
-    std::unique_ptr<ShaderProgramInternal> createProgram(std::vector<uint32_t> &&modules);
-    void destroyProgram(std::unique_ptr<ShaderProgramInternal> &ptr);
-    //void reloadPrograms();
 
     vk::Device getDevice() const { return vkDevice; } /*make device global*/
   };
