+++ etna/source/ShaderProgram.cpp
@@ -5,154 +5,6 @@
 
 namespace etna
 {
-  void DescriptorSetInfo::addResource(const vk::DescriptorSetLayoutBinding &binding)
-  {
-    if (binding.binding > MAX_DESCRIPTOR_BINDINGS)
-      throw std::runtime_error {"DescriptorSetInfo: Binding out of MAX_DESCRIPTOR_BINDINGS range"};
-
-    if (usedBindings.test(binding.binding))
-    {
-      auto &src = bindings[binding.binding];
-      if (src.descriptorType != binding.descriptorType
-        || src.descriptorCount != binding.descriptorCount)
-      {
-        throw std::runtime_error {"DescriptorSetInfo: incompatible bindings"};
-      }
-
-      src.stageFlags |= binding.stageFlags;
-      return;
-    }
-    
-    usedBindings.set(binding.binding);
-    bindings[binding.binding] = binding;
-
-    if (binding.binding + 1 > maxUsedBinding)
-      maxUsedBinding = binding.binding + 1;
-    if (binding.descriptorType == vk::DescriptorType::eUniformBufferDynamic || binding.descriptorType == vk::DescriptorType::eStorageBufferDynamic)
-      dynOffsets++;
-  }
-  
-  void DescriptorSetInfo::clear()
-  {
-    maxUsedBinding = 0;
-    dynOffsets = 0;
-    usedBindings.reset();
-    for (auto &binding : bindings)
-      binding = vk::DescriptorSetLayoutBinding {};
-  }
-
-  void DescriptorSetInfo::parseShader(vk::ShaderStageFlagBits stage, const SpvReflectDescriptorSet &spv)
-  {
-    for (uint32_t i = 0u; i < spv.binding_count; i++)
-    {
-      const auto &spvBinding = *spv.bindings[i];
-      
-      vk::DescriptorSetLayoutBinding apiBinding {};
-      apiBinding.descriptorCount = 1;
-
-      for (uint32_t i = 0; i < spvBinding.array.dims_count; i++) {
-        apiBinding.descriptorCount *= spvBinding.array.dims[i];
-      }
-
-      apiBinding.descriptorType = static_cast<vk::DescriptorType>(spvBinding.descriptor_type);
-      apiBinding.stageFlags = stage;
-      apiBinding.pImmutableSamplers = nullptr;
-      apiBinding.binding = spvBinding.binding;
-      addResource(apiBinding);
-    }
-  }
-  
-  void DescriptorSetInfo::merge(const DescriptorSetInfo &info)
-  {
-    for (uint32_t binding = 0; binding < info.maxUsedBinding; binding++)
-    {
-      if (!info.usedBindings.test(binding))
-        continue;
-      addResource(info.bindings[binding]);
-    }
-  }
-
-  bool DescriptorSetInfo::operator==(const DescriptorSetInfo &rhs) const
-  {
-    if (maxUsedBinding != rhs.maxUsedBinding)
-      return false;
-    if (usedBindings != rhs.usedBindings)
-      return false;
-    
-    for (uint32_t i = 0; i < maxUsedBinding; i++)
-    {
-      if (bindings[i] != rhs.bindings[i])
-        return false;
-    }
-
-    return true;
-  }
-
-  vk::DescriptorSetLayout DescriptorSetInfo::createVkLayout(vk::Device device) const
-  {
-    std::vector<vk::DescriptorSetLayoutBinding> apiBindings;
-    for (uint32_t i = 0; i < maxUsedBinding; i++)
-    {
-      if (!usedBindings.test(i))
-        continue;
-      apiBindings.push_back(bindings[i]);
-    }
-
-    vk::DescriptorSetLayoutCreateInfo info {};
-    info.setBindings(apiBindings);
-    
-    return device.createDescriptorSetLayout(info);
-  }
-
-  template <typename T>
-  inline void hash_combine(std::size_t &s, const T &v)
-  {
-    std::hash<T> h;
-    s ^= h(v) + 0x9e3779b9 + (s<< 6) + (s>> 2); 
-  }
-
-  std::size_t DescriptorSetLayoutHash::operator()(const DescriptorSetInfo &res) const
-  {
-    size_t hash = 0;
-
-    for (uint32_t i = 0; i < res.maxUsedBinding; i++)
-    {
-      if (!res.usedBindings.test(i))
-        continue;
-        
-      hash_combine(hash, res.bindings[i].binding);
-      hash_combine(hash, static_cast<uint32_t>(res.bindings[i].descriptorType));
-      hash_combine(hash, res.bindings[i].descriptorCount);
-      hash_combine(hash, static_cast<uint32_t>(res.bindings[i].stageFlags));
-    }
-
-    return hash;
-  }
-
-  DescriptorLayoutId DescriptorSetLayoutCache::registerLayout(vk::Device device, const DescriptorSetInfo &info)
-  {
-    auto it = map.find(info);
-    if (it != map.end())
-      return it->second;
-    
-    DescriptorLayoutId id = descriptors.size();
-    map.insert({info, id});
-    descriptors.push_back(info);
-    vkLayouts.push_back(info.createVkLayout(device));
-    return id;
-  }
-  
-  void DescriptorSetLayoutCache::clear(vk::Device device)
-  {
-    for (auto layout : vkLayouts) {
-      device.destroyDescriptorSetLayout(layout);
-    }
-
-    map.clear();
-    descriptors.clear();
-    vkLayouts.clear();
-  }
-
   ShaderModule::ShaderModule(vk::Device device, const std::string &shader_path)
     : path {shader_path}
   {
@@ -269,69 +121,6 @@ namespace etna
     return modId;
   }
 
-  std::unique_ptr<ShaderProgramManager::ShaderProgramInternal> ShaderProgramManager::createProgram(std::vector<uint32_t> &&modules)
-  {
-    std::unique_ptr<ShaderProgramInternal> prog;
-    prog.reset(new ShaderProgramInternal {});
-    prog->moduleIds = std::move(modules);
-    prog->usedDescriptors.reset();
-    prog->pushConst = vk::PushConstantRange {};
-
-    std::array<DescriptorSetInfo, MAX_PROGRAM_DESCRIPTORS> dstDescriptors;
-
-    for (auto id : prog->moduleIds)
-    {
-      auto &shaderMod = *shaderModules[id];
-
-      if (shaderMod.getPushConstUsed()) //merge push constants
-      {
-        auto pushConst = shaderMod.getPushConst();
-        if (prog->pushConst.size == 0u)
-        {
-          prog->pushConst = pushConst;
-        }
-        else {
-          if (prog->pushConst.size != pushConst.size)
-            throw std::runtime_error {"ShaderProgram: not compatible push constant blocks"};
-          prog->pushConst.stageFlags |= pushConst.stageFlags;
-        }
-      }
-
-      const auto &resources = shaderMod.getResources(); //merge descriptors
-      for (auto &desc : resources)
-      {
-        if (desc.first >= MAX_PROGRAM_DESCRIPTORS)
-          throw std::runtime_error {"ShaderProgram: Descriptor set out of range"};
-
-        prog->usedDescriptors.set(desc.first);
-        dstDescriptors[desc.first].merge(desc.second);
-      }
-    }
-
-    std::vector<vk::DescriptorSetLayout> vkLayouts;
-
-    for (uint32_t i = 0; i < MAX_PROGRAM_DESCRIPTORS; i++)
-    {
-      if (!prog->usedDescriptors.test(i))
-        continue;
-      prog->descriptorIds[i] = desciptorLayoutCache.registerLayout(getDevice(), dstDescriptors[i]);
-      vkLayouts.push_back(desciptorLayoutCache.getVkLayout(prog->descriptorIds[i]));
-    }
-
-    vk::PipelineLayoutCreateInfo info {};
-    info.setSetLayouts(vkLayouts);
-
-    if (prog->pushConst.size)
-    {
-      info.setPPushConstantRanges(&prog->pushConst);
-      info.setPushConstantRangeCount(1u);
-    }
-
-    prog->progLayout = getDevice().createPipelineLayout(info);    
-
-    return prog;
-  }
-
   static void validate_program_shaders(const std::vector<vk::ShaderStageFlagBits> &stages) {
     auto supportedShaders = 
       vk::ShaderStageFlagBits::eVertex|
@@ -381,7 +170,8 @@ namespace etna
     validate_program_shaders(stages);
 
     ShaderProgramId progId = programs.size();
-    programs.push_back(createProgram(std::move(moduleIds)));
+    programs.emplace_back(new ShaderProgramInternal {std::move(moduleIds)});
+    programs[progId]->reload(*this);
     programNames[name] = progId;
     return progId;
   }
@@ -394,11 +184,72 @@ namespace etna
     return it->second;
   }
 
-  void ShaderProgramManager::destroyProgram(std::unique_ptr<ShaderProgramInternal> &ptr)
+  void ShaderProgramManager::ShaderProgramInternal::reload(ShaderProgramManager &manager)
   {
-    if (ptr->progLayout)
-      getDevice().destroyPipelineLayout(ptr->progLayout);
-    ptr.reset(nullptr);
+    destroy(manager);
+
+    std::array<DescriptorSetInfo, MAX_PROGRAM_DESCRIPTORS> dstDescriptors;
+
+    for (auto id : moduleIds)
+    {
+      auto &shaderMod = manager.getModule(id);
+
+      if (shaderMod.getPushConst().size) //merge push constants
+      {
+        auto modPushConst = shaderMod.getPushConst();
+        if (pushConst.size == 0u)
+        {
+          pushConst = modPushConst;
+        }
+        else {
+          if (pushConst.size != modPushConst.size)
+            throw std::runtime_error {"ShaderProgram: not compatible push constant blocks"};
+          pushConst.stageFlags |= modPushConst.stageFlags;
+        }
+      }
+
+      const auto &resources = shaderMod.getResources(); //merge descriptors
+      for (auto &desc : resources)
+      {
+        if (desc.first >= MAX_PROGRAM_DESCRIPTORS)
+          throw std::runtime_error {"ShaderProgram: Descriptor set out of range"};
+
+        usedDescriptors.set(desc.first);
+        dstDescriptors[desc.first].merge(desc.second);
+      }
+    }
+
+    std::vector<vk::DescriptorSetLayout> vkLayouts;
+
+    for (uint32_t i = 0; i < MAX_PROGRAM_DESCRIPTORS; i++)
+    {
+      if (!usedDescriptors.test(i))
+        continue;
+      auto res = manager.desciptorLayoutCache.get(manager.getDevice(), dstDescriptors[i]);
+      descriptorIds[i] = res.first;
+      vkLayouts.push_back(res.second);
+    }
+
+    vk::PipelineLayoutCreateInfo info {};
+    info.setSetLayouts(vkLayouts);
+
+    if (pushConst.size)
+    {
+      info.setPPushConstantRanges(&pushConst);
+      info.setPushConstantRangeCount(1u);
+    }
+
+    progLayout = manager.getDevice().createPipelineLayout(info);
+  }
+  
+  void ShaderProgramManager::ShaderProgramInternal::destroy(ShaderProgramManager &manager)
+  {
+    if (progLayout)
+      manager.getDevice().destroyPipelineLayout(progLayout);
+    progLayout = nullptr;
+
+    usedDescriptors.reset();
+    pushConst = vk::PushConstantRange {};
   }
 
   void ShaderProgramManager::reloadPrograms()
@@ -412,9 +263,7 @@ namespace etna
 
     for (auto &progPtr : programs)
     {
-      auto moduleIds = std::move(progPtr->moduleIds);
-      destroyProgram(progPtr);
-      progPtr = createProgram(std::move(moduleIds));
+      progPtr->reload(*this);
     }
   }
   
@@ -422,7 +271,7 @@ namespace etna
   {
     programNames.clear();
     for (auto &progPtr : programs)
-      destroyProgram(progPtr);
+      progPtr->destroy(*this);
     programs.clear();
 
     shaderModuleNames.clear();
